Allow caller scope to control the selection
This lets us get a caller owned state that existed when we started,
as well as watch when it changes.

We still broadcast event when the user tries to select something. If
the caller has not indicated they don't want to control the selection
then we automatically select the object the user clicked on.